Make frontend read dark mode pref from backend
The JS frontend now reads from the preference stored in the Crystal
backend to update the localStorage on page load, and checks the
preference stored in the backend when updating the mode. This should
fix the issues noted in PR #601 regarding the frontend and backend
prefs not being synchronized properly.