info message for youtube-import (#740)
since signing up through the big button on the welcome message leads to
/data_control, it needs a message on what to do next. the most common
option here is importing subscriptions from youtube, hence it was moved
to the top.
the rather inconspicuous link to google's support pages was replaced by
a infobox that instructs the user to click a link to download a file and
upload it on this page.