Store continuation so that it can be used on the next matching request.
This is done because the continuation created for videos sorted by
'oldest' doesn't work after the first 30 videos. The same 30 videos are
returned again.

The only way to get the next 30 videos, and onwards, is to use the
continuation returned in the initial API call. Storing the returned
continuation in the db saves having to request each page from 1 to the
currently wanted page each time a page other than the first is wanted.